move MetadataItem to entity
