--- packages/web/src/components/entity-page/EntityActionButton.tsx
@@ -1,8 +1,0 @@
-import { Button, ButtonProps } from '@audius/harmony'
-
-type CollectionActionButtonProps = ButtonProps
-
-export const EntityActionButton = (props: CollectionActionButtonProps) => {
-  const { children, ...other } = props
-  return <Button {...other}>{children}</Button>
-}
